- Add null checks to stop crashing

+++ src/android/Capture.java
@@ -348,8 +348,8 @@ public class Capture extends CordovaPlugin {
         //set video save path in the intent
         takeVideoIntent.putExtra(MediaStore.EXTRA_OUTPUT, videoUri);
         takeVideoIntent.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
-
-
+        
+        
         if (Build.VERSION.SDK_INT > 7) {
             takeVideoIntent.putExtra("android.intent.extra.durationLimit", req.duration);
             takeVideoIntent.putExtra("android.intent.extra.videoQuality", req.quality);
@@ -392,6 +392,7 @@ public class Capture extends CordovaPlugin {
      */
     public void onActivityResult(int requestCode, int resultCode, final Intent intent) {
         final Request req = pendingRequests.get(requestCode);
+
         // Result received okay
         if (resultCode == Activity.RESULT_OK) {
             Runnable processActivityResult = new Runnable() {
@@ -402,10 +403,12 @@ public class Capture extends CordovaPlugin {
                             onAudioActivityResult(req, intent);
                             break;
                         case CAPTURE_IMAGE_OR_VIDEO:
-
                             if (checkURIResource(videoUri)) {
+                                Log.d("pastrami", "video result");
                                 onVideoActivityResult(req);
-                            } else {
+                            }
+                            else {
+                                Log.d("pastrami", "image result");
                                 onImageActivityResult(req);
                             }
                             break;
@@ -419,25 +422,23 @@ public class Capture extends CordovaPlugin {
         else if (resultCode == Activity.RESULT_CANCELED) {
             CleanUpEmptyVideo();
             // If we have partial results send them back to the user
-            if (req.results.length() > 0 && req.results != null) {
+            if (req != null && req.results != null && req.results.length() > 0) {
                 pendingRequests.resolveWithSuccess(req);
             }
             // user canceled the action
-            else {
+            else if(req != null){
                 pendingRequests.resolveWithFailure(req, createErrorObject(CAPTURE_NO_MEDIA_FILES, "Canceled."));
             }
         }
         // If something else
         else {
             CleanUpEmptyVideo();
-            // If we have partial results send them back to the user
-            if (req.results.length() > 0) {
-
+            // If we have partial results send them back to the userI
+            if (req != null && req.results != null && req.results.length() > 0) {
                 pendingRequests.resolveWithSuccess(req);
             }
             // something bad happened
-            else {
-
+            else if(req != null){
                 pendingRequests.resolveWithFailure(req, createErrorObject(CAPTURE_NO_MEDIA_FILES, "Did not complete!"));
             }
         }
@@ -455,7 +456,7 @@ public class Capture extends CordovaPlugin {
                 @Override
                 public void run() {
                     File videoFile = webView.getResourceApi().mapUriToFile(videoUri);
-                    if (videoFile.exists()) {
+                    if (videoFile != null && videoFile.exists()) {
                         videoFile.delete();
                     }
                 }
